requirements.txt
# Core dependencies
pyyaml>=6.0.1
pillow>=10.0.0
numpy>=1.24.0

# Data platform
picsellia>=6.0.0

# ML Framework
ultralytics>=8.1.0  # YOLOv11
torch>=2.0.0
torchvision>=0.15.0

# MLflow and tracking
mlflow>=2.18.0
psycopg2-binary>=2.9.9  # PostgreSQL for MLflow
boto3>=1.34.0  # S3 for MLflow artifacts

# Preprocessing
deface>=1.5.0  # Face anonymization
opencv-python>=4.8.0
opencv-python-headless>=4.8.0

# Serving
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
aiofiles>=23.2.1

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.1.0
